unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@557054 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/types/resources/StringResource.java
@@ -144,9 +144,8 @@ public class StringResource extends Resource {
      *         compatibility with java.io.File), or UNKNOWN_SIZE if not known.
      */
     public synchronized long getSize() {
-        return isReference()
-            ? ((Resource) getCheckedRef()).getSize()
-            : (long) getContent().length();
+        return isReference() ? ((Resource) getCheckedRef()).getSize()
+                : getContent().length();
     }
 
     /**
